requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
distro==1.8.0
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
multidict==6.0.4
numpy==1.26.1
openai==0.28.0
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.6.0
yarl==1.9.2
